FINAL BUILD 1.0 !!! :)

+++ Lab_4/structs.h
@@ -4,7 +4,7 @@
 struct Actions {
 	int operations = 0;
 	int swaps = 0;
-	int time = 0;
+	float time = 0;
 };
